+++ tests/test_subscription.py
@@ -13,9 +13,13 @@ from django.contrib.auth import get_user_model
 from django.test import TestCase
 from django.utils import timezone
 from mock import patch
+from stripe.error import InvalidRequestError
 
 from djstripe.models import Customer, Subscription, Plan
-from tests import FAKE_SUBSCRIPTION, FAKE_PLAN, FAKE_CUSTOMER, FAKE_PLAN_II, datetime_to_unix
+from tests import (
+    datetime_to_unix, FAKE_CUSTOMER, FAKE_PLAN, FAKE_PLAN_II,
+    FAKE_SUBSCRIPTION, FAKE_SUBSCRIPTION_CANCELED
+)
 
 
 class SubscriptionTest(TestCase):
@@ -55,6 +59,7 @@ class SubscriptionTest(TestCase):
         self.assertTrue(subscription.is_status_current())
         self.assertTrue(subscription.is_status_temporarily_current())
         self.assertTrue(subscription.is_valid())
+        self.assertTrue(subscription in self.customer.active_subscriptions)
         self.assertTrue(self.customer.has_active_subscription())
         self.assertTrue(self.customer.has_any_active_subscription())
 
@@ -69,6 +74,7 @@ class SubscriptionTest(TestCase):
         self.assertTrue(subscription.is_status_current())
         self.assertFalse(subscription.is_status_temporarily_current())
         self.assertTrue(subscription.is_valid())
+        self.assertTrue(subscription in self.customer.active_subscriptions)
         self.assertTrue(self.customer.has_active_subscription())
         self.assertTrue(self.customer.has_any_active_subscription())
 
@@ -84,6 +90,7 @@ class SubscriptionTest(TestCase):
         self.assertFalse(subscription.is_status_current())
         self.assertFalse(subscription.is_status_temporarily_current())
         self.assertFalse(subscription.is_valid())
+        self.assertFalse(subscription in self.customer.active_subscriptions)
         self.assertFalse(self.customer.has_active_subscription())
         self.assertFalse(self.customer.has_any_active_subscription())
 
@@ -97,6 +104,8 @@ class SubscriptionTest(TestCase):
         subscription_retrieve_mock.return_value = subscription_fake
 
         subscription = Subscription.sync_from_stripe_data(subscription_fake)
+        self.assertFalse(subscription in self.customer.active_subscriptions)
+        self.assertEquals(self.customer.active_subscriptions.count(), 0)
 
         delta = timezone.timedelta(days=30)
         extended_subscription = subscription.extend(delta)
@@ -195,6 +204,7 @@ class SubscriptionTest(TestCase):
         subscription_retrieve_mock.return_value = canceled_subscription_fake  # retrieve().delete()
 
         self.assertTrue(self.customer.has_active_subscription())
+        self.assertEquals(self.customer.active_subscriptions.count(), 1)
         self.assertTrue(self.customer.has_any_active_subscription())
 
         new_subscription = subscription.cancel(at_period_end=False)
@@ -203,6 +213,7 @@ class SubscriptionTest(TestCase):
         self.assertEqual(False, new_subscription.cancel_at_period_end)
         self.assertEqual(new_subscription.canceled_at, new_subscription.ended_at)
         self.assertFalse(new_subscription.is_valid())
+        self.assertFalse(new_subscription in self.customer.active_subscriptions)
         self.assertFalse(self.customer.has_active_subscription())
         self.assertFalse(self.customer.has_any_active_subscription())
 
@@ -224,8 +235,12 @@ class SubscriptionTest(TestCase):
 
         self.assertTrue(self.customer.has_active_subscription())
         self.assertTrue(self.customer.has_any_active_subscription())
+        self.assertEquals(self.customer.active_subscriptions.count(), 1)
+        self.assertTrue(subscription in self.customer.active_subscriptions)
 
         new_subscription = subscription.cancel(at_period_end=True)
+        self.assertEquals(self.customer.active_subscriptions.count(), 1)
+        self.assertTrue(new_subscription in self.customer.active_subscriptions)
 
         self.assertEqual(Subscription.STATUS_ACTIVE, new_subscription.status)
         self.assertEqual(True, new_subscription.cancel_at_period_end)
@@ -289,3 +304,25 @@ class SubscriptionTest(TestCase):
         subscription_reactivate_fake = deepcopy(FAKE_SUBSCRIPTION)
         reactivated_subscription = Subscription.sync_from_stripe_data(subscription_reactivate_fake)
         self.assertEqual(reactivated_subscription.cancel_at_period_end, False)
+
+    @patch("djstripe.stripe_objects.StripeSubscription._api_delete")
+    @patch("stripe.Subscription.retrieve", return_value=deepcopy(FAKE_SUBSCRIPTION_CANCELED))
+    def test_cancel_already_canceled(self, subscription_retrieve_mock, subscription_delete_mock):
+        subscription_delete_mock.side_effect = InvalidRequestError("No such subscription: sub_xxxx", "blah")
+
+        subscription_fake = deepcopy(FAKE_SUBSCRIPTION)
+        subscription = Subscription.sync_from_stripe_data(subscription_fake)
+
+        self.assertEqual(Subscription.objects.filter(status="canceled").count(), 0)
+        subscription.cancel()
+        self.assertEqual(Subscription.objects.filter(status="canceled").count(), 1)
+
+    @patch("djstripe.stripe_objects.StripeSubscription._api_delete")
+    def test_cancel_error_in_cancel(self, subscription_delete_mock):
+        subscription_delete_mock.side_effect = InvalidRequestError("Unexpected error", "blah")
+
+        subscription_fake = deepcopy(FAKE_SUBSCRIPTION)
+        subscription = Subscription.sync_from_stripe_data(subscription_fake)
+
+        with self.assertRaises(InvalidRequestError):
+            subscription.cancel()